refactor: convert the `clickOutsideOrEscape` action to an attachment
Additionally, bump `prettier-plugin-svelte` to version 3.4.0 so svelte attachments are known to the linter (are still highlighted as errors in WebStorm)
