+++ package.json
@@ -68,7 +68,7 @@
         "mode-watcher": "^1.1.0",
         "postcss": "^8.5.6",
         "prettier": "^3.6.2",
-        "prettier-plugin-svelte": "^3.3.3",
+        "prettier-plugin-svelte": "3.4.0",
         "prettier-plugin-tailwindcss": "^0.6.14",
         "protobuf-fieldmask": "^2.0.0",
         "svelte": "^5.45.2",
